new files for more readibility ( closes #50)
